Fit eDNA Model

Fit eDNA model

eDNA_lm(
  formula,
  data,
  std_curve_alpha,
  std_curve_beta,
  upper_Cq = 40,
  n_chain = 4L,
  iters = 1000L,
  verbose = FALSE,
  prior_intercept = normal(location = -15, scale = 10),
  priors = normal(),
  Cq_error_type = "fixed",
  sink_file = tempfile(),
  ...
)

eDNA_lmer(
  formula,
  data,
  std_curve_alpha,
  std_curve_beta,
  upper_Cq = 40,
  n_chain = 4L,
  iters = 500L,
  verbose = FALSE,
  prior_intercept = normal(location = -15, scale = 10),
  priors = normal(),
  Cq_error_type = "fixed",
  sink_file = tempfile(),
  ...
)

Arguments

formula
a formula, specifying the relationship between the predictors and the latent variable eDNA concentration.

data
data.frame, with the response and predictors

std_curve_alpha
the alpha value for the formula for converting between log(eDNA concentration) and CQ value

std_curve_beta
the beta value for the formula for converting between log(eDNA concentration) and CQ value

upper_Cq
numeric, the upper limit on CQ detection. Any value of log(concentration) which would result in a value greater than this limit is instead recorded as the limit.

n_chain
integer, the number of chains to use. Please note that less than two is not recommended.

iters
integer, the number of iterations per chain

verbose
logical, when TRUE output from rstan::sampling is written to the console.

prior_intercept
named list such as created by rstanarm::normal. The list must contain elements named "location" and "scale", which are the location and scale for a normal prior over the intercept. Ignored when the intercept is omitted in the model formula.

priors
named list such as created by rstanarm::normal. The list must contain elements named "location" and "scale", which are the location and scale for a normal prior over the betas, and "autoscale". If a single value is provided, this value will be repeated for each beta. If autoscale = TRUE, the scale of the priors is scaled by the sd of the predictors similar to rstanarm handles them.

Cq_error_type
either "fixed" or "varying", specifying if measurement error is assumed to be the same for all values of CQ, or if it increases or decreases as CQ increases.

sink_file
character, a file to write the console output to if verbose = FALSE, by default writes to tempfile()

...
additional arguments passed to sampling

Value

S4 object, with the following slots:

ln_conc

matrix, the posterior samples for the latent variable, eDNA concentration

Cq_star

matrix, the posterior prediction for the observed response

betas

array, the posterior estimates for the betas for the linear model

sigma_Cq

array, the posterior estimates for the measurement error of CQ

formula

formula, the original formula used in the model

x

data.frame, the model matrix used in the model

std_curve_alpha

numeric, the std. curve value used

std_curve_beta

numeric, the std. curve value used

upper_Cq

numeric, the upper limit for observed CQ used

stanfit

stanfit, the original results from rstan::sampling

Details

These functions fit a Bayesian latent variable model to data collected from a eDNA sampling experiment. These data have a few particular characteristics that justify using a specialized model. More details on these characteristics and the model structure, please refer to the "Getting Started" vignette for the artemis package.

There are two different modeling functions in the artemis package, eDNA_lm and eDNA_lmer. eDNA_lm is for fitting a fixed effects model, while eDNA_lmer is for fitting a mixed or random effects model. Both models are fit using the rstan::sampling function, which uses a Hamiltonian Monte Carlo algorithm to estimate parameters for the model. Users are encouraged to refer to the documentation for Stan and RStan at https://mc-stan.org/users/documentation/ for details about how models are fit.

Diagnosing warning and error messages

The models have been written in Stan with key focus on robustness and speed. However, it is possible that users might encounter issues. Typically, these issues will be highlighted by warning messages, which will be displayed regardless of the value of verbose. Often times, these warnings can be resolved by increasing the number of iterations that the HMC algorithm runs by specifying iters to be a larger value. This should be the first action attempted, as increasing the iters increases both the warm-up and sampling iterations. If users continue to have issues, additional control arguments can be passed to rstan::sampling via the ... argument.

Examples

## Fixed effect model
ans = eDNA_lm(Cq ~ Distance, eDNA_data,
              std_curve_alpha = 21.2, std_curve_beta = -1.5)

#> Model executable is up to date!

#> Running MCMC with 4 chain(s) on 1 core(s)...
#> 
#> Running ./eDNA_omni 'id=1' random 'seed=346871191' data \
#>   'file=/tmp/Rtmpe8oEkC/standata-10dfe79a4e35.json' output \
#>   'file=/tmp/Rtmpe8oEkC/eDNA_omni-202006231344-1-b8aec1.csv' 'method=sample' \
#>   'save_warmup=0' 'algorithm=hmc' 'engine=nuts' adapt 'engaged=1'

#> Chain 1 Rejecting initial value:

#> Chain 1   Gradient evaluated at the initial value is not finite.

#> Chain 1   Stan can't start sampling from this initial value.

#> Chain 1 Rejecting initial value:

#> Chain 1   Gradient evaluated at the initial value is not finite.

#> Chain 1   Stan can't start sampling from this initial value.

#> Chain 1 Rejecting initial value:

#> Chain 1   Gradient evaluated at the initial value is not finite.

#> Chain 1   Stan can't start sampling from this initial value.

#> Chain 1 Rejecting initial value:

#> Chain 1   Gradient evaluated at the initial value is not finite.

#> Chain 1   Stan can't start sampling from this initial value.

#> Chain 1 Rejecting initial value:

#> Chain 1   Gradient evaluated at the initial value is not finite.

#> Chain 1   Stan can't start sampling from this initial value.

#> Chain 1 Rejecting initial value:

#> Chain 1   Gradient evaluated at the initial value is not finite.

#> Chain 1   Stan can't start sampling from this initial value.

#> Chain 1 Rejecting initial value:

#> Chain 1   Gradient evaluated at the initial value is not finite.

#> Chain 1   Stan can't start sampling from this initial value.

#> Chain 1 Gradient evaluation took 8.8e-05 seconds

#> Chain 1 1000 transitions using 10 leapfrog steps per transition would take 0.88 seconds.

#> Chain 1 Adjust your expectations accordingly!

#> Chain 1 Iteration:    1 / 2000 [  0%]  (Warmup) 
#> Chain 1 Iteration:  100 / 2000 [  5%]  (Warmup) 
#> Chain 1 Iteration:  200 / 2000 [ 10%]  (Warmup) 
#> Chain 1 Iteration:  300 / 2000 [ 15%]  (Warmup) 
#> Chain 1 Iteration:  400 / 2000 [ 20%]  (Warmup) 
#> Chain 1 Iteration:  500 / 2000 [ 25%]  (Warmup) 
#> Chain 1 Iteration:  600 / 2000 [ 30%]  (Warmup) 
#> Chain 1 Iteration:  700 / 2000 [ 35%]  (Warmup) 
#> Chain 1 Iteration:  800 / 2000 [ 40%]  (Warmup) 
#> Chain 1 Iteration:  900 / 2000 [ 45%]  (Warmup) 
#> Chain 1 Iteration: 1000 / 2000 [ 50%]  (Warmup) 
#> Chain 1 Iteration: 1001 / 2000 [ 50%]  (Sampling) 
#> Chain 1 Iteration: 1100 / 2000 [ 55%]  (Sampling) 
#> Chain 1 Iteration: 1200 / 2000 [ 60%]  (Sampling) 
#> Chain 1 Iteration: 1300 / 2000 [ 65%]  (Sampling) 
#> Chain 1 Iteration: 1400 / 2000 [ 70%]  (Sampling) 
#> Chain 1 Iteration: 1500 / 2000 [ 75%]  (Sampling) 
#> Chain 1 Iteration: 1600 / 2000 [ 80%]  (Sampling) 
#> Chain 1 Iteration: 1700 / 2000 [ 85%]  (Sampling) 
#> Chain 1 Iteration: 1800 / 2000 [ 90%]  (Sampling) 
#> Chain 1 Iteration: 1900 / 2000 [ 95%]  (Sampling) 
#> Chain 1 Iteration: 2000 / 2000 [100%]  (Sampling) 
#> Chain 1 finished in 7.7 seconds.
#> Running ./eDNA_omni 'id=2' random 'seed=432744511' data \
#>   'file=/tmp/Rtmpe8oEkC/standata-10dfe79a4e35.json' output \
#>   'file=/tmp/Rtmpe8oEkC/eDNA_omni-202006231344-2-b8aec1.csv' 'method=sample' \
#>   'save_warmup=0' 'algorithm=hmc' 'engine=nuts' adapt 'engaged=1'

#> Chain 2 Rejecting initial value:

#> Chain 2   Gradient evaluated at the initial value is not finite.

#> Chain 2   Stan can't start sampling from this initial value.

#> Chain 2 Rejecting initial value:

#> Chain 2   Gradient evaluated at the initial value is not finite.

#> Chain 2   Stan can't start sampling from this initial value.

#> Chain 2 Rejecting initial value:

#> Chain 2   Gradient evaluated at the initial value is not finite.

#> Chain 2   Stan can't start sampling from this initial value.

#> Chain 2 Rejecting initial value:

#> Chain 2   Gradient evaluated at the initial value is not finite.

#> Chain 2   Stan can't start sampling from this initial value.

#> Chain 2 Rejecting initial value:

#> Chain 2   Gradient evaluated at the initial value is not finite.

#> Chain 2   Stan can't start sampling from this initial value.

#> Chain 2 Rejecting initial value:

#> Chain 2   Gradient evaluated at the initial value is not finite.

#> Chain 2   Stan can't start sampling from this initial value.

#> Chain 2 Gradient evaluation took 0.000104 seconds

#> Chain 2 1000 transitions using 10 leapfrog steps per transition would take 1.04 seconds.

#> Chain 2 Adjust your expectations accordingly!

#> Chain 2 Iteration:    1 / 2000 [  0%]  (Warmup) 
#> Chain 2 Iteration:  100 / 2000 [  5%]  (Warmup) 
#> Chain 2 Iteration:  200 / 2000 [ 10%]  (Warmup) 
#> Chain 2 Iteration:  300 / 2000 [ 15%]  (Warmup) 
#> Chain 2 Iteration:  400 / 2000 [ 20%]  (Warmup) 
#> Chain 2 Iteration:  500 / 2000 [ 25%]  (Warmup) 
#> Chain 2 Iteration:  600 / 2000 [ 30%]  (Warmup) 
#> Chain 2 Iteration:  700 / 2000 [ 35%]  (Warmup) 
#> Chain 2 Iteration:  800 / 2000 [ 40%]  (Warmup) 
#> Chain 2 Iteration:  900 / 2000 [ 45%]  (Warmup) 
#> Chain 2 Iteration: 1000 / 2000 [ 50%]  (Warmup) 
#> Chain 2 Iteration: 1001 / 2000 [ 50%]  (Sampling) 
#> Chain 2 Iteration: 1100 / 2000 [ 55%]  (Sampling) 
#> Chain 2 Iteration: 1200 / 2000 [ 60%]  (Sampling) 
#> Chain 2 Iteration: 1300 / 2000 [ 65%]  (Sampling) 
#> Chain 2 Iteration: 1400 / 2000 [ 70%]  (Sampling) 
#> Chain 2 Iteration: 1500 / 2000 [ 75%]  (Sampling) 
#> Chain 2 Iteration: 1600 / 2000 [ 80%]  (Sampling) 
#> Chain 2 Iteration: 1700 / 2000 [ 85%]  (Sampling) 
#> Chain 2 Iteration: 1800 / 2000 [ 90%]  (Sampling) 
#> Chain 2 Iteration: 1900 / 2000 [ 95%]  (Sampling) 
#> Chain 2 Iteration: 2000 / 2000 [100%]  (Sampling) 
#> Chain 2 finished in 8.1 seconds.
#> Running ./eDNA_omni 'id=3' random 'seed=31749990' data \
#>   'file=/tmp/Rtmpe8oEkC/standata-10dfe79a4e35.json' output \
#>   'file=/tmp/Rtmpe8oEkC/eDNA_omni-202006231344-3-b8aec1.csv' 'method=sample' \
#>   'save_warmup=0' 'algorithm=hmc' 'engine=nuts' adapt 'engaged=1'

#> Chain 3 Rejecting initial value:

#> Chain 3   Gradient evaluated at the initial value is not finite.

#> Chain 3   Stan can't start sampling from this initial value.

#> Chain 3 Gradient evaluation took 0.000229 seconds

#> Chain 3 1000 transitions using 10 leapfrog steps per transition would take 2.29 seconds.

#> Chain 3 Adjust your expectations accordingly!

#> Chain 3 Iteration:    1 / 2000 [  0%]  (Warmup) 

#> Chain 3 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:

#> Chain 3 Exception: normal_lpdf: Scale parameter[1] is 0, but must be > 0! (in '/tmp/RtmpGK0qHi/model-e18d6ef39c68.stan', line 231, column 1 to line 234, column 40)

#> Chain 3 If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,

#> Chain 3 but if this warning occurs often then your model may be either severely ill-conditioned or misspecified.

#> Chain 3 

#> Chain 3 Iteration:  100 / 2000 [  5%]  (Warmup) 
#> Chain 3 Iteration:  200 / 2000 [ 10%]  (Warmup) 
#> Chain 3 Iteration:  300 / 2000 [ 15%]  (Warmup) 
#> Chain 3 Iteration:  400 / 2000 [ 20%]  (Warmup) 
#> Chain 3 Iteration:  500 / 2000 [ 25%]  (Warmup) 
#> Chain 3 Iteration:  600 / 2000 [ 30%]  (Warmup) 
#> Chain 3 Iteration:  700 / 2000 [ 35%]  (Warmup) 
#> Chain 3 Iteration:  800 / 2000 [ 40%]  (Warmup) 
#> Chain 3 Iteration:  900 / 2000 [ 45%]  (Warmup) 
#> Chain 3 Iteration: 1000 / 2000 [ 50%]  (Warmup) 
#> Chain 3 Iteration: 1001 / 2000 [ 50%]  (Sampling) 
#> Chain 3 Iteration: 1100 / 2000 [ 55%]  (Sampling) 
#> Chain 3 Iteration: 1200 / 2000 [ 60%]  (Sampling) 
#> Chain 3 Iteration: 1300 / 2000 [ 65%]  (Sampling) 
#> Chain 3 Iteration: 1400 / 2000 [ 70%]  (Sampling) 
#> Chain 3 Iteration: 1500 / 2000 [ 75%]  (Sampling) 
#> Chain 3 Iteration: 1600 / 2000 [ 80%]  (Sampling) 
#> Chain 3 Iteration: 1700 / 2000 [ 85%]  (Sampling) 
#> Chain 3 Iteration: 1800 / 2000 [ 90%]  (Sampling) 
#> Chain 3 Iteration: 1900 / 2000 [ 95%]  (Sampling) 
#> Chain 3 Iteration: 2000 / 2000 [100%]  (Sampling) 
#> Chain 3 finished in 8.4 seconds.
#> Running ./eDNA_omni 'id=4' random 'seed=2137934383' data \
#>   'file=/tmp/Rtmpe8oEkC/standata-10dfe79a4e35.json' output \
#>   'file=/tmp/Rtmpe8oEkC/eDNA_omni-202006231344-4-b8aec1.csv' 'method=sample' \
#>   'save_warmup=0' 'algorithm=hmc' 'engine=nuts' adapt 'engaged=1'
#> Chain 4 Iteration:    1 / 2000 [  0%]  (Warmup) 
#> Chain 4 Iteration:  100 / 2000 [  5%]  (Warmup) 
#> Chain 4 Iteration:  200 / 2000 [ 10%]  (Warmup) 
#> Chain 4 Iteration:  300 / 2000 [ 15%]  (Warmup) 
#> Chain 4 Iteration:  400 / 2000 [ 20%]  (Warmup) 
#> Chain 4 Iteration:  500 / 2000 [ 25%]  (Warmup) 
#> Chain 4 Iteration:  600 / 2000 [ 30%]  (Warmup) 
#> Chain 4 Iteration:  700 / 2000 [ 35%]  (Warmup) 
#> Chain 4 Iteration:  800 / 2000 [ 40%]  (Warmup) 
#> Chain 4 Iteration:  900 / 2000 [ 45%]  (Warmup) 
#> Chain 4 Iteration: 1000 / 2000 [ 50%]  (Warmup) 
#> Chain 4 Iteration: 1001 / 2000 [ 50%]  (Sampling) 
#> Chain 4 Iteration: 1100 / 2000 [ 55%]  (Sampling) 
#> Chain 4 Iteration: 1200 / 2000 [ 60%]  (Sampling) 
#> Chain 4 Iteration: 1300 / 2000 [ 65%]  (Sampling) 
#> Chain 4 Iteration: 1400 / 2000 [ 70%]  (Sampling) 
#> Chain 4 Iteration: 1500 / 2000 [ 75%]  (Sampling) 
#> Chain 4 Iteration: 1600 / 2000 [ 80%]  (Sampling) 
#> Chain 4 Iteration: 1700 / 2000 [ 85%]  (Sampling) 
#> Chain 4 Iteration: 1800 / 2000 [ 90%]  (Sampling) 
#> Chain 4 Iteration: 1900 / 2000 [ 95%]  (Sampling) 
#> Chain 4 Iteration: 2000 / 2000 [100%]  (Sampling) 
#> Chain 4 finished in 7.9 seconds.
#> 
#> All 4 chains finished successfully.
#> Mean chain execution time: 8.0 seconds.
#> Total execution time: 32.2 seconds.

## Mixed-effect model
## This takes a while to run
if (FALSE) {
ans2 = eDNA_lmer(Cq ~ Distance + (1|SampleID), eDNA_data,
                 std_curve_alpha = 21.2, std_curve_beta = -1.5)

}