eDNA model fit results

An S4 object holding the results of a model fit by eDNA_lm*

Slots

ln_conc

matrix, the estimated latent variable, log(eDNA concentration)

Cq_star

matrix, the predicted CQ value for each obs

betas

matrix, the posterior estimate for each beta in the model

sigma_ln_eDNA

the estimated measurement error on ln_eDNA

formula

the formula used to fit the model

x

the model.matrix used to fit the model

std_curve_alpha

the alpha for the std. curve conversion formual used

std_curve_beta

the beta for the std. curve conversion formula used

upper_Cq

the upper limit for CQ

stanfit

the result from rstan::sampling

random_x

data.frame of the grouping variables used

random_sd

the estimated stdev. of each of the random effects