Plot p(detect)

Plot the p(detect eDNA)

# S3 method for eDNA_p_detect
plot(x, y, probs = c(0.025, 0.975), ylim = c(0,
  1), point_size = 2.5, n_breaks = 5, error_width = rel(1.5), ...)

Arguments

x
object of class eDNA_p_detect

y
ignored

probs
probabilities for error bars when a model fit is supplied

ylim
limits on the y axis

point_size
point size for the estimated p(detect)

n_breaks
passed to pretty

error_width
the width of the ends of errorbars

...
additional args passed to plot.stanfit

Value

a ggplot object

Details

This allows the visualization of the probability of detection producted by est_p_detect. If the object includes an interval estimate, errorbars are produced with those intervals.