Fit eDNA Model

Fit eDNA model

eDNA_lm(formula, data, std_curve_alpha, std_curve_beta, upper_Cq = 40,
  n_chain = 4L, iters = 1000L, verbose = FALSE,
  sink_file = tempfile(), ...)

eDNA_lmer(formula, data, std_curve_alpha, std_curve_beta, upper_Cq = 40,
  n_chain = 4L, iters = 500L, verbose = FALSE,
  sink_file = tempfile(), ...)

Arguments

formula
a formula, specifying the relationship between the predictors and the latent variable eDNA concentration.

data
data.frame, with the response and predictors

std_curve_alpha
the alpha value for the formula for converting between log(eDNA concentration) and CQ value

std_curve_beta
the beta value for the formula for converting between log(eDNA concentration) and CQ value

upper_Cq
numeric, the upper limit on CQ detection. Any value of log(concentration) which would result in a value greater than this limit is instead recorded as the limit.

n_chain
integer, the number of chains to use. Please note that less than two is not recommended.

iters
integer, the number of iterations per chain

verbose
logical, when TRUE output from rstan::sampling is written to the console.

sink_file
character, a file to write the console output to if verbose = FALSE, by default writes to tempfile()

...
additional arguments passed to sampling

Value

S4 object, with the following slots:

ln_conc

matrix, the posterior samples for the latent variable, eDNA concentration

Cq_star

matrix, the posterior prediction for the observed response

betas

array, the posterior estimates for the betas for the linear model

sigma_Cq

array, the posterior estimates for the measurement error of CQ

formula

formula, the original formula used in the model

x

data.frame, the model matrix used in the model

std_curve_alpha

numeric, the std. curve value used

std_curve_beta

numeric, the std. curve value used

upper_Cq

numeric, the upper limit for observed CQ used

stanfit

stanfit, the original results from rstan::sampling

Details

These functions fit a Bayesian latent variable model to data collected from a eDNA sampling experiment. These data have a few particular characteristics that justify using a specialized model. More details on these characteristics and the model structure, please refer to the "Getting Started" vignette for the artemis package.

There are two different modeling functions in the artemis package, eDNA_lm and eDNA_lmer. eDNA_lm is for fitting a fixed effects model, while eDNA_lmer is for fitting a mixed or random effects model. Both models are fit using the rstan::sampling function, which uses a Hamiltonian Monte Carlo algorithm to estimate parameters for the model. Users are encouraged to refer to the documentation for Stan and RStan at https://mc-stan.org/users/documentation/ for details about how models are fit.

Diagnosing warning and error messages

The models have been written in Stan with key focus on robustness and speed. However, it is possible that users might encounter issues. Typically, these issues will be highlighted by warning messages, which will be displayed regardless of the value of verbose. Often times, these warnings can be resolved by increasing the number of iterations that the HMC algorithm runs by specifying iters to be a larger value. This should be the first action attempted, as increasing the iters increases both the warm-up and sampling iterations. If users continue to have issues, additional control arguments can be passed to rstan::sampling via the ... argument.

Examples

## Fixed effect model
ans = eDNA_lm(Cq ~ Distance, eDNA_data,
              std_curve_alpha = 21.2, std_curve_beta = -1.5)

## Mixed-effect model
## This takes a while to run
if (FALSE) {
ans2 = eDNA_lmer(Cq ~ Distance + (1|SampleID), eDNA_data,
                 std_curve_alpha = 21.2, std_curve_beta = -1.5)

}