Fit eDNA Model

Fit eDNA model

eDNA_lm(
  formula,
  data,
  std_curve_alpha,
  std_curve_beta,
  upper_Cq = 40,
  prior_intercept = normal(location = -15, scale = 10),
  priors = normal(),
  ...
)

eDNA_lmer(
  formula,
  data,
  std_curve_alpha,
  std_curve_beta,
  upper_Cq = 40,
  prior_intercept = normal(location = -15, scale = 10),
  priors = normal(),
  prior_random_variance = exponential(),
  ...
)

eDNA_zinf_lm(
  formula,
  data,
  std_curve_alpha,
  std_curve_beta,
  upper_Cq = 40,
  probability_zero = 0.08,
  prior_intercept = normal(location = -15, scale = 10),
  priors = normal(),
  ...
)

eDNA_zinf_lmer(
  formula,
  data,
  std_curve_alpha,
  std_curve_beta,
  upper_Cq = 40,
  probability_zero = 0.08,
  prior_intercept = normal(location = -15, scale = 10),
  priors = normal(),
  prior_random_variance = exponential(),
  ...
)

Arguments

formula
a formula, specifying the relationship between the predictors and the latent variable eDNA concentration.

data
data.frame, with the response and predictors

std_curve_alpha
the alpha value for the formula for converting between log(eDNA concentration) and CQ value

std_curve_beta
the beta value for the formula for converting between log(eDNA concentration) and CQ value

upper_Cq
numeric, the upper limit on CQ detection. Any value of log(concentration) which would result in a value greater than this limit is instead recorded as the limit.

prior_intercept
named list such as created by rstanarm::normal. The list must contain elements named "location" and "scale", which are the location and scale for a normal prior over the intercept. Ignored when the intercept is omitted in the model formula.

priors
named list such as created by rstanarm::normal. The list must contain elements named "location" and "scale", which are the location and scale for a normal prior over the betas, and "autoscale". If a single value is provided, this value will be repeated for each beta. If autoscale = TRUE, the scale of the priors is scaled by the sd of the predictors similar to rstanarm handles them.

...
additional arguments passed to sampling

prior_random_variance
the prior on variance of the random effects. Defaults to exponential distribution with rate 1.

probability_zero
numeric, between 0 and 1. The probability of a non-detection from a source other than low concentration of eDNA, e.g. a filter failure. Defaults to 8 was the estimated p(zero) from a daily sampling experiment.

Value

S4 object, with the following slots:

ln_conc

matrix, the posterior samples for the latent variable, eDNA concentration

Cq_star

matrix, the posterior prediction for the observed response

betas

array, the posterior estimates for the betas for the linear model

sigma_ln_eDNA

array, the posterior estimates for the measurement error of ln_eDNA

formula

formula, the original formula used in the model

x

data.frame, the model matrix used in the model

std_curve_alpha

numeric, the std. curve value used

std_curve_beta

numeric, the std. curve value used

upper_Cq

numeric, the upper limit for observed CQ used

stanfit

stanfit, the original results from rstan::sampling

Details

These functions fit a Bayesian latent variable model to data collected from a eDNA sampling experiment. These data have a few particular characteristics that justify using a specialized model. More details on these characteristics and the model structure, please refer to the "Getting Started" vignette for the artemis package.

There are four different modeling functions in the artemis package, eDNA_lm, eDNA_lmer, eDNA_zinf_lm, eDNA_zinf_lmer. eDNA_lm is for fitting a fixed effects model, while eDNA_lmer is for fitting a mixed or random effects model. The *_zinf versions implement a zero-inflated version of their respective lm function. All models are fit using the rstan::sampling function, which uses a Hamiltonian Monte Carlo algorithm to estimate parameters for the model. Users are encouraged to refer to the documentation for Stan and RStan at https://mc-stan.org/users/documentation/ for details about how models are fit.

Diagnosing warning and error messages

The models have been written in Stan with key focus on robustness and speed. However, it is possible that users might encounter issues. Typically, these issues will be highlighted by warning messages coming from rstan::sampling. Often times, these warnings can be resolved by increasing the number of iterations that the HMC algorithm runs by specifying iters to be a larger value. This should be the first action attempted, as increasing the iters increases both the warm-up and sampling iterations. If users continue to have issues, additional control arguments can be passed to rstan::sampling via the ... argument.

Author

Matt Espe

Examples

## Fixed effect model
ans = eDNA_lm(Cq ~ Distance_m, eDNA_data,
              std_curve_alpha = 21.2, std_curve_beta = -1.5)

#> 
#> SAMPLING FOR MODEL 'eDNA_lm' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 5.7e-05 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0.57 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.152665 seconds (Warm-up)
#> Chain 1:                0.103261 seconds (Sampling)
#> Chain 1:                0.255926 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'eDNA_lm' NOW (CHAIN 2).
#> Chain 2: Rejecting initial value:
#> Chain 2:   Log probability evaluates to log(0), i.e. negative infinity.
#> Chain 2:   Stan can't start sampling from this initial value.
#> Chain 2: 
#> Chain 2: Gradient evaluation took 1.7e-05 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 0.17 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 0.147476 seconds (Warm-up)
#> Chain 2:                0.120369 seconds (Sampling)
#> Chain 2:                0.267845 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'eDNA_lm' NOW (CHAIN 3).
#> Chain 3: Rejecting initial value:
#> Chain 3:   Log probability evaluates to log(0), i.e. negative infinity.
#> Chain 3:   Stan can't start sampling from this initial value.
#> Chain 3: 
#> Chain 3: Gradient evaluation took 1.6e-05 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 0.16 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 0.137531 seconds (Warm-up)
#> Chain 3:                0.111835 seconds (Sampling)
#> Chain 3:                0.249366 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'eDNA_lm' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 2.1e-05 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 0.21 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 0.146543 seconds (Warm-up)
#> Chain 4:                0.112858 seconds (Sampling)
#> Chain 4:                0.259401 seconds (Total)
#> Chain 4: 

# \donttest{
## Mixed-effect model
## This takes a while to run
ans2 = eDNA_lmer(Cq ~ Distance_m + (1|FilterID), eDNA_data,
                 std_curve_alpha = 21.2, std_curve_beta = -1.5)

#> 
#> SAMPLING FOR MODEL 'eDNA_lmer' NOW (CHAIN 1).
#> Chain 1: Rejecting initial value:
#> Chain 1:   Log probability evaluates to log(0), i.e. negative infinity.
#> Chain 1:   Stan can't start sampling from this initial value.
#> Chain 1: 
#> Chain 1: Gradient evaluation took 8e-05 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0.8 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 3.8895 seconds (Warm-up)
#> Chain 1:                1.25765 seconds (Sampling)
#> Chain 1:                5.14715 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'eDNA_lmer' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 5.3e-05 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 0.53 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 3.8652 seconds (Warm-up)
#> Chain 2:                1.21872 seconds (Sampling)
#> Chain 2:                5.08391 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'eDNA_lmer' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 4.7e-05 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 0.47 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 4.13785 seconds (Warm-up)
#> Chain 3:                1.44993 seconds (Sampling)
#> Chain 3:                5.58778 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'eDNA_lmer' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 5.3e-05 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 0.53 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 3.73793 seconds (Warm-up)
#> Chain 4:                1.21047 seconds (Sampling)
#> Chain 4:                4.9484 seconds (Total)
#> Chain 4: 

# }