Estimate the probability of detection

Estimate the probability of detection

est_p_detect(
  variable_levels,
  betas,
  Cq_sd,
  std_curve_alpha,
  std_curve_beta,
  n_rep = 1:12,
  model_fit = NULL,
  upper_Cq = 40
)

Arguments

variable_levels
numeric vector, with each element corresponding to the condition to estimate the probability of detection.

betas
numeric vector, the effect sizes for each of the variable level

Cq_sd
numeric, the measurement error on CQ. If a model_fit is provided and this is missing, the estimated sd(CQ) from the model will be used.

std_curve_alpha
the alpha for the std. curve formula for conversion between log(concentration) and CQ

std_curve_beta
the alpha for the std. curve formula for conversion between log(concentration) and CQ

n_rep
the number of replicate measurements at the levels specified

model_fit
optional, a model fit from eDNA_lm or eDNA_lmer. If this is provided, an estimate derived from the posterior estimates of beta is calculated.

upper_Cq
the upper limit on detection

Value

object of class "eDNA_p_detect" with the estimates of the probability of detection for the variable levels provided.

Details

This function estimates the probability of getting a positive detection for an eDNA survey given a set of predictors. This can be useful when trying to take the estimates from a preliminary study and use those estimates to inform the deployment of future sampling schemes. The function assumes that you have either an idea of the effects of the various predictors, for example from a previous study, or a fit model with estimates of the effect sizes.

This function takes one circumstance at a time, and calculates the range of outcomes given a number of repeated sampling attempts. The probability calculated is the probability of getting at least one positive detection. For details on the underlying model and assumptions for this calculation, please refer to the package vignette.

Notes on random effects

This function deals with random effects in two different ways. First, when we desire to see the probability of detection for a specific instance of a random effect, users can specify the random effect as just another effect by specifying the random effect = 1 in the variable list, and then the size of the random effect. However, when users wish to estimate the probability of detection in cases where random effects are generated from a distribution of random effects, this can be accomplished by adding the standard deviation of the random effect to the Cq_sd. This takes advantage of the fact that random effects are just another source of variation, and that sum of random normal distributions is itself a random normal distribution.

Examples

est_p_detect(variable_levels = c(Intercept = 1, Distance = 100, Volume = 20),
             betas = c(Intercept = -10.5, Distance = -0.05, Volume = 0.001),
             Cq_sd = 1, std_curve_alpha = 21.2, std_curve_beta = -1.5,
             n_rep = 1:12)

#> Variable levels: 
#> Intercept  Distance    Volume 
#>         1       100        20 
#> 
#>    n_reps     p_detect
#> 1       1 4.935045e-06
#> 2       2 9.870066e-06
#> 3       3 1.480506e-05
#> 4       4 1.974003e-05
#> 5       5 2.467498e-05
#> 6       6 2.960991e-05
#> 7       7 3.454480e-05
#> 8       8 3.947968e-05
#> 9       9 4.441453e-05
#> 10     10 4.934935e-05
#> 11     11 5.428416e-05
#> 12     12 5.921893e-05